Moved tests from old TxBuilderTest files to relevant new test files
