+++ core-test/src/test/scala/org/bitcoins/core/wallet/builder/RawTxSignerTest.scala
@@ -1,108 +1,73 @@
 package org.bitcoins.core.wallet.builder
 
-import org.bitcoins.core.config.RegTest
-import org.bitcoins.core.crypto._
-import org.bitcoins.core.currency._
+import org.bitcoins.core.crypto.{
+  BaseTxSigComponent,
+  WitnessTxSigComponentP2SH,
+  WitnessTxSigComponentRaw
+}
+import org.bitcoins.core.currency.{Bitcoins, CurrencyUnits, Satoshis}
 import org.bitcoins.core.number.UInt32
 import org.bitcoins.core.policy.Policy
-import org.bitcoins.core.protocol.script._
-import org.bitcoins.core.protocol.transaction._
+import org.bitcoins.core.protocol.script.{
+  CLTVScriptPubKey,
+  CSVScriptPubKey,
+  ConditionalScriptPubKey,
+  EmptyScriptPubKey,
+  MultiSignatureScriptPubKey,
+  NonStandardScriptPubKey,
+  P2PKHScriptPubKey,
+  P2PKScriptPubKey,
+  P2PKWithTimeoutScriptPubKey,
+  P2SHScriptPubKey,
+  P2SHScriptSignature,
+  P2WSHWitnessV0,
+  UnassignedWitnessScriptPubKey,
+  WitnessCommitment,
+  WitnessScriptPubKey,
+  WitnessScriptPubKeyV0
+}
+import org.bitcoins.core.protocol.transaction.{
+  BaseTransaction,
+  Transaction,
+  TransactionConstants,
+  TransactionInput,
+  TransactionOutPoint,
+  TransactionOutput,
+  WitnessTransaction
+}
 import org.bitcoins.core.script.PreExecutionScriptProgram
 import org.bitcoins.core.script.constant.ScriptNumber
 import org.bitcoins.core.script.crypto.HashType
 import org.bitcoins.core.script.interpreter.ScriptInterpreter
 import org.bitcoins.core.wallet.fee.{SatoshisPerByte, SatoshisPerVirtualByte}
-import org.bitcoins.core.wallet.utxo._
+import org.bitcoins.core.wallet.utxo.{
+  ConditionalPath,
+  InputInfo,
+  InputSigningInfo,
+  LockTimeInputInfo,
+  ScriptSignatureParams
+}
 import org.bitcoins.crypto.{DoubleSha256DigestBE, ECPrivateKey}
 import org.bitcoins.testkit.Implicits._
 import org.bitcoins.testkit.core.gen.{CreditingTxGen, ScriptGenerators}
 import org.bitcoins.testkit.util.BitcoinSAsyncTest
 
-class BitcoinTxBuilderTest extends BitcoinSAsyncTest {
-  val tc: TransactionConstants.type = TransactionConstants
-  val (spk, privKey) = ScriptGenerators.p2pkhScriptPubKey.sampleSome
-
+class RawTxSignerTest extends BitcoinSAsyncTest {
   implicit override val generatorDrivenConfig: PropertyCheckConfiguration =
     generatorDrivenConfigNewCode
 
-  behavior of "BitcoinTxBuilder"
+  behavior of "RawTxSigner"
 
-  it must "failed to build a transaction that mints money out of thin air" in {
+  private val (spk, privKey) = ScriptGenerators.p2pkhScriptPubKey.sampleSome
 
+  it should "fail a transaction when the user invariants fail" in {
     val creditingOutput = TransactionOutput(CurrencyUnits.zero, spk)
     val destinations =
       Vector(TransactionOutput(Satoshis.one, EmptyScriptPubKey))
-    val creditingTx = BaseTransaction(version = tc.validLockVersion,
-                                      inputs = Nil,
-                                      outputs = Vector(creditingOutput),
-                                      lockTime = tc.lockTime)
-    val outPoint = TransactionOutPoint(creditingTx.txId, UInt32.zero)
-    val utxo = ScriptSignatureParams(
-      InputInfo(
-        outPoint = outPoint,
-        output = creditingOutput,
-        redeemScriptOpt = None,
-        scriptWitnessOpt = None,
-        conditionalPath = ConditionalPath.NoCondition,
-        hashPreImages = Vector(privKey.publicKey)
-      ),
-      signer = privKey,
-      hashType = HashType.sigHashAll
-    )
-    val utxos = Vector(utxo)
-    val feeUnit = SatoshisPerVirtualByte(Satoshis.one)
-
-    recoverToSucceededIf[IllegalArgumentException] {
-      val utxF =
-        NonInteractiveWithChangeFinalizer.txFrom(outputs = destinations,
-                                                 utxos = utxos,
-                                                 feeRate = feeUnit,
-                                                 changeSPK = EmptyScriptPubKey)
-
-      utxF.flatMap(utx => RawTxSigner.sign(utx, utxos, feeUnit))
-    }
-  }
-
-  it must "fail to build a transaction when we pass in a negative fee rate" in {
-    val creditingOutput = TransactionOutput(CurrencyUnits.zero, spk)
-    val destinations =
-      Vector(TransactionOutput(Satoshis.one, EmptyScriptPubKey))
-    val creditingTx = BaseTransaction(version = tc.validLockVersion,
-                                      inputs = Nil,
-                                      outputs = Vector(creditingOutput),
-                                      lockTime = tc.lockTime)
-    val outPoint = TransactionOutPoint(creditingTx.txId, UInt32.zero)
-    val utxo = ScriptSignatureParams(
-      InputInfo(
-        outPoint = outPoint,
-        output = creditingOutput,
-        redeemScriptOpt = None,
-        scriptWitnessOpt = None,
-        conditionalPath = ConditionalPath.NoCondition,
-        hashPreImages = Vector(privKey.publicKey)
-      ),
-      signer = privKey,
-      hashType = HashType.sigHashAll
-    )
-    val utxos = Vector(utxo)
-    val feeUnit = SatoshisPerVirtualByte(Satoshis(-1))
-
-    recoverToSucceededIf[IllegalArgumentException] {
-      NonInteractiveWithChangeFinalizer.txFrom(outputs = destinations,
-                                               utxos = utxos,
-                                               feeRate = feeUnit,
-                                               changeSPK = EmptyScriptPubKey)
-    }
-  }
-
-  it must "fail a transaction when the user invariants fail" in {
-    val creditingOutput = TransactionOutput(CurrencyUnits.zero, spk)
-    val destinations =
-      Vector(TransactionOutput(Satoshis.one, EmptyScriptPubKey))
-    val creditingTx = BaseTransaction(tc.validLockVersion,
+    val creditingTx = BaseTransaction(TransactionConstants.validLockVersion,
                                       Nil,
                                       Vector(creditingOutput),
-                                      tc.lockTime)
+                                      TransactionConstants.lockTime)
     val outPoint = TransactionOutPoint(creditingTx.txId, UInt32.zero)
     val utxo =
       ScriptSignatureParams(
@@ -132,87 +97,15 @@ class BitcoinTxBuilderTest extends BitcoinSAsyncTest {
     }
   }
 
-  it must "fail to build a tx if you have the wrong redeemscript" in {
-    val p2sh = P2SHScriptPubKey(spk)
-    val creditingOutput = TransactionOutput(CurrencyUnits.zero, p2sh)
-    val creditingTx = BaseTransaction(version = tc.validLockVersion,
-                                      inputs = Nil,
-                                      outputs = Vector(creditingOutput),
-                                      lockTime = tc.lockTime)
-    val outPoint = TransactionOutPoint(creditingTx.txId, UInt32.zero)
-    val inputInfo = InputInfo(
-      outPoint = outPoint,
-      output = creditingOutput,
-      redeemScriptOpt = Some(EmptyScriptPubKey),
-      scriptWitnessOpt = None,
-      conditionalPath = ConditionalPath.NoCondition
-    )
-
-    assertThrows[RuntimeException] {
-      ScriptSignatureParams(
-        inputInfo = inputInfo,
-        signer = privKey,
-        hashType = HashType.sigHashAll
-      )
-    }
-
-    assertThrows[RuntimeException] {
-      ECSignatureParams(
-        inputInfo = inputInfo,
-        signer = privKey,
-        hashType = HashType.sigHashAll
-      )
-    }
-  }
-
-  it must "fail to build a tx if you have the wrong script witness" in {
-    val p2wsh = P2WSHWitnessSPKV0(spk)
-    val creditingOutput = TransactionOutput(CurrencyUnits.zero, p2wsh)
-    val creditingTx = BaseTransaction(tc.validLockVersion,
-                                      Nil,
-                                      Vector(creditingOutput),
-                                      tc.lockTime)
-    val outPoint = TransactionOutPoint(creditingTx.txId, UInt32.zero)
-    assertThrows[IllegalArgumentException] {
-      ScriptSignatureParams(
-        InputInfo(
-          outPoint = outPoint,
-          output = creditingOutput,
-          redeemScriptOpt = None,
-          scriptWitnessOpt = Some(P2WSHWitnessV0(EmptyScriptPubKey)),
-          conditionalPath = ConditionalPath.NoCondition,
-          hashPreImages = Vector(privKey.publicKey)
-        ),
-        privKey,
-        HashType.sigHashAll
-      )
-    }
-
-    assertThrows[IllegalArgumentException] {
-      ECSignatureParams(
-        InputInfo(
-          outPoint = outPoint,
-          output = creditingOutput,
-          redeemScriptOpt = None,
-          scriptWitnessOpt = Some(P2WSHWitnessV0(EmptyScriptPubKey)),
-          conditionalPath = ConditionalPath.NoCondition,
-          hashPreImages = Vector(privKey.publicKey)
-        ),
-        privKey,
-        HashType.sigHashAll
-      )
-    }
-  }
-
-  it must "fail to sign a p2pkh if we don't pass in the public key" in {
+  it should "fail to sign a p2pkh if we don't pass in the public key" in {
     val p2pkh = P2PKHScriptPubKey(privKey.publicKey)
     val creditingOutput = TransactionOutput(CurrencyUnits.zero, p2pkh)
     val destinations =
       Vector(TransactionOutput(Satoshis.one, EmptyScriptPubKey))
-    val creditingTx = BaseTransaction(tc.validLockVersion,
+    val creditingTx = BaseTransaction(TransactionConstants.validLockVersion,
                                       Nil,
                                       Vector(creditingOutput),
-                                      tc.lockTime)
+                                      TransactionConstants.lockTime)
     val outPoint = TransactionOutPoint(creditingTx.txId, UInt32.zero)
     val utxo = ScriptSignatureParams(
       InputInfo(
@@ -240,15 +133,16 @@ class BitcoinTxBuilderTest extends BitcoinSAsyncTest {
     }
   }
 
-  it must "fail to sign a p2pkh if we pass in the wrong public key" in {
+  it should "fail to sign a p2pkh if we pass in the wrong public key" in {
     val p2pkh = P2PKHScriptPubKey(privKey.publicKey)
     val creditingOutput = TransactionOutput(CurrencyUnits.zero, p2pkh)
     val destinations =
       Vector(TransactionOutput(Satoshis.one, EmptyScriptPubKey))
-    val creditingTx = BaseTransaction(version = tc.validLockVersion,
+    val creditingTx = BaseTransaction(version =
+                                        TransactionConstants.validLockVersion,
                                       inputs = Nil,
                                       outputs = Vector(creditingOutput),
-                                      lockTime = tc.lockTime)
+                                      lockTime = TransactionConstants.lockTime)
     val outPoint =
       TransactionOutPoint(txId = creditingTx.txId, vout = UInt32.zero)
     val utxo = ScriptSignatureParams(
@@ -277,81 +171,7 @@ class BitcoinTxBuilderTest extends BitcoinSAsyncTest {
     }
   }
 
-  it must "fail to sign a p2wpkh if we don't pass in the public key" in {
-    val p2wpkh = P2WPKHWitnessSPKV0(pubKey = privKey.publicKey)
-    val creditingOutput =
-      TransactionOutput(value = CurrencyUnits.zero, scriptPubKey = p2wpkh)
-    val creditingTx = BaseTransaction(version = tc.validLockVersion,
-                                      inputs = Nil,
-                                      outputs = Vector(creditingOutput),
-                                      lockTime = tc.lockTime)
-    val outPoint =
-      TransactionOutPoint(txId = creditingTx.txId, vout = UInt32.zero)
-    val inputInfo = InputInfo(
-      outPoint = outPoint,
-      output = creditingOutput,
-      redeemScriptOpt = None,
-      scriptWitnessOpt = Some(P2WSHWitnessV0(EmptyScriptPubKey)),
-      conditionalPath = ConditionalPath.NoCondition
-    )
-
-    assertThrows[IllegalArgumentException] {
-      ScriptSignatureParams(
-        inputInfo = inputInfo,
-        signer = privKey,
-        hashType = HashType.sigHashAll
-      )
-    }
-
-    assertThrows[IllegalArgumentException] {
-      ECSignatureParams(
-        inputInfo = inputInfo,
-        signer = privKey,
-        hashType = HashType.sigHashAll
-      )
-    }
-  }
-
-  it must "fail to sign a p2wpkh if we pass in the wrong public key" in {
-    val p2wpkh = P2WPKHWitnessSPKV0(privKey.publicKey)
-    val creditingOutput = TransactionOutput(CurrencyUnits.zero, p2wpkh)
-    val creditingTx = BaseTransaction(tc.validLockVersion,
-                                      Nil,
-                                      Vector(creditingOutput),
-                                      tc.lockTime)
-    val outPoint = TransactionOutPoint(creditingTx.txId, UInt32.zero)
-    assertThrows[IllegalArgumentException] {
-      ScriptSignatureParams(
-        InputInfo(
-          outPoint = outPoint,
-          output = creditingOutput,
-          redeemScriptOpt = None,
-          scriptWitnessOpt = Some(P2WSHWitnessV0(EmptyScriptPubKey)),
-          conditionalPath = ConditionalPath.NoCondition,
-          hashPreImages = Vector(privKey.publicKey)
-        ),
-        privKey,
-        HashType.sigHashAll
-      )
-    }
-
-    assertThrows[IllegalArgumentException] {
-      ECSignatureParams(
-        InputInfo(
-          outPoint = outPoint,
-          output = creditingOutput,
-          redeemScriptOpt = None,
-          scriptWitnessOpt = Some(P2WSHWitnessV0(EmptyScriptPubKey)),
-          conditionalPath = ConditionalPath.NoCondition,
-          hashPreImages = Vector(privKey.publicKey)
-        ),
-        privKey,
-        HashType.sigHashAll
-      )
-    }
-  }
-
-  it must "succeed to sign a cltv spk that uses a second-based locktime" in {
+  it should "succeed to sign a cltv spk that uses a second-based locktime" in {
     val fundingPrivKey = ECPrivateKey.freshPrivateKey
 
     val lockTime = System.currentTimeMillis / 1000
@@ -388,7 +208,7 @@ class BitcoinTxBuilderTest extends BitcoinSAsyncTest {
       .map(tx => assert(tx.lockTime == UInt32(lockTime)))
   }
 
-  it must "succeed to sign a cltv spk that uses a block height locktime" in {
+  it should "succeed to sign a cltv spk that uses a block height locktime" in {
     val fundingPrivKey = ECPrivateKey.freshPrivateKey
 
     val lockTime = 1000
@@ -425,7 +245,7 @@ class BitcoinTxBuilderTest extends BitcoinSAsyncTest {
       .map(tx => assert(tx.lockTime == UInt32(lockTime)))
   }
 
-  it must "fail to sign a cltv spk that uses both a second-based and a block height locktime" in {
+  it should "fail to sign a cltv spk that uses both a second-based and a block height locktime" in {
     val fundingPrivKey1 = ECPrivateKey.freshPrivateKey
     val fundingPrivKey2 = ECPrivateKey.freshPrivateKey
 
@@ -475,35 +295,6 @@ class BitcoinTxBuilderTest extends BitcoinSAsyncTest {
     recoverToSucceededIf[IllegalArgumentException](
       utxF.flatMap(utx => RawTxSigner.sign(utx, utxos, feeRate))
     )
-  }
-
-  it must "fail to construct a tx given an UnassignedSegwitNativeUTXOSpendingInfo" in {
-    val outPoint = TransactionOutPoint(DoubleSha256DigestBE.empty, UInt32.zero)
-    val privKey = ECPrivateKey.freshPrivateKey
-    val pubKey = privKey.publicKey
-
-    val spendingInfo =
-      ScriptSignatureParams(
-        UnassignedSegwitNativeInputInfo(
-          outPoint = outPoint,
-          amount = Bitcoins.one + CurrencyUnits.oneMBTC,
-          scriptPubKey = P2WPKHWitnessSPKV0(pubKey),
-          scriptWitness = P2WPKHWitnessV0(pubKey),
-          conditionalPath = ConditionalPath.NoCondition,
-          Vector(pubKey)
-        ),
-        signers = Vector(privKey),
-        hashType = HashType.sigHashAll
-      )
-
-    recoverToSucceededIf[UnsupportedOperationException] {
-      NonInteractiveWithChangeFinalizer.txFrom(
-        Vector(TransactionOutput(Bitcoins.one, EmptyScriptPubKey)),
-        Vector(spendingInfo),
-        SatoshisPerVirtualByte(Satoshis.one),
-        EmptyScriptPubKey
-      )
-    }
   }
 
   def verifyScript(
@@ -564,7 +355,7 @@ class BitcoinTxBuilderTest extends BitcoinSAsyncTest {
     ScriptInterpreter.runAllVerify(programs)
   }
 
-  it must "sign a mix of spks in a tx and then have it verified" in {
+  it should "sign a mix of spks in a tx and then have it verified" in {
     forAllAsync(CreditingTxGen.inputsAndOutputs(),
                 ScriptGenerators.scriptPubKey) {
       case ((creditingTxsInfo, destinations), (changeSPK, _)) =>
@@ -583,7 +374,7 @@ class BitcoinTxBuilderTest extends BitcoinSAsyncTest {
     }
   }
 
-  it must "sign a mix of p2sh/p2wsh in a tx and then have it verified" in {
+  it should "sign a mix of p2sh/p2wsh in a tx and then have it verified" in {
     forAllAsync(CreditingTxGen.inputsAndOutputs(CreditingTxGen.nestedOutputs),
                 ScriptGenerators.scriptPubKey) {
       case ((creditingTxsInfo, destinations), (changeSPK, _)) =>
